Refactor all documentation-related functions into their own module
